Keyple 0.8.0 - Fixed build on Windows/MSVC

+++ component/keyple-common/src/main/System.h
@@ -68,7 +68,7 @@ public:
 
     static unsigned long long currentTimeMillis()
     {
-        return nanoTime() / pow(10, 6);
+        return (unsigned long long)(nanoTime() / pow(10, 6));
     }
 };
